Add files via upload
Fixed the slipping animation, turned out it was because I was not calculating the correct angle using the actual coordinates of the body and the foot. Instead, I was just using the touchdown angle, which was incorrect as it would just be incorrect and cause the slipping due to the xtd not being constant in the math. Now to fix up some small tidbits of animation and move onto the controller!